updated comments and minor code changes
