--- author-a-map/src/main/res/drawable/border.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <solid android:color="#ffffff" />
-    <stroke
-        android:width="1dip"
-        android:color="#388E3C" />
-</shape>
