--- author-a-map/src/main/res/layout/auth_view.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <com.esri.android.oauth.OAuthView xmlns:android="http://schemas.android.com/apk/res/android"
-        android:id="@+id/oauthview"
-        clientid=""
-        portalurl="https://arcgis.com"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"></com.esri.android.oauth.OAuthView>
-</FrameLayout>
-
