--- author-a-map/src/main/res/layout/list_item.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <TextView
-        android:id="@+id/textview"
-        android:layout_width="match_parent"
-        android:layout_height="65dp"
-        android:gravity="center|fill"
-        android:textAlignment="center"
-        android:textStyle="bold"></TextView>
-
-</LinearLayout>
